A big update to catch up on changes to Students
Basically, we'd revamped how we do Users in the database but hadn't updated iOS stuff.

Also, because this seemed like a good opportunity to move to something a bit more thought out, I've landed on using Moya as an abstraction layer for all of this stuff. Should make API stuff much more clear about what it's doing

Also, explicitly forcing serialization of the User's ID property in the database code. Not sure why we wouldn't do this anyways, it's nice to have that so we can navigate to specific user pages for updates / deletions