Reworking some things; adding remainder of models
This commit broadly focuses on adding the models we've currently defined in our Django backend, but you'll see a couple minor changes squeeze in too.

Most notably, I realized there was a better way to implement the `toDictionary()` function, and I fixed a bug in the `searchForStudents` call where it wasn't using the appropriate request encoding.