Fixed bugs with passing functions
This was a kinda weird one to fix; With the instrument service, for example, we were getting errors when adding/removing instruments of "this.list isn't a function"...

It would still *work*, but it did require a refresh to see the added/removed artifacts. It turns out that in this situation (and others), we had to wrap the function in a lambda
so that instead of passing "this.update", we pass "() => this.update()". I don't know enough about JavaScript nor am I bothered to look up why this behavior is different. Oh well