Refactored the login method
Instead of using a state variable `currentUserToken` to pass data between the concatMaps, I passed it via a map call.
The tap which created a student object (if it existed) was moved so it was attached to the get-student API call, which absolved us of the need to check if "thing" has an m_number.