The Ensemble Dashboard now retains the currently opened ensemble
HOWEVER: There is a bug that if you open another ensemble without closing another one, the route parameter is set to none-selected, which isn't the desired behavior. This seems to be because the opened event of the new ensemble is fired _before_ the closed event of the previous ensemble, thus the final state is that the query is closed.

When an ensemble is opened, its ID is added to the current URL as an optional route parameter. The component observes changes to this route parameter to expand one of the ensembles at a time.