Rework Invoices and more GraphQL
Instead of the old Invoice/Maintenance/Purchase hierarchy, we moved over to a more generic and flexible schema. Invoice is no longer abstract and has a Many-to-Many with Asset, through LineItem. Every LineItem also has a type (purchase or maintenance, for now), a cost, and notes. Assets can still only have one purchase line-item.

These changes were reflected in the GraphQL api, but have not yet been reflected in the REST api.