Added a Cost History component
We still need to implement features to add estimated resale value and cost to replace, not to mention styling, but all the bones for "display" are here.

We should probably add an ability to get the "details" of lineItem, such as its receipt. We also need to add interaction and wire up the buttons.